Fix handling of bg class id in SegmentationMapOnImage
Changed so that bool and float input arrays to
SegmentationMapOnImage implicitly mean that any spatial
location without a value >threshold in all array channels
is counted as background (i.e. all channels zero-like
at a pixel = background). For integer maps, the background
class is explicitly provided. It may be changed in
get_arr_int() and the drawing routines.