Reduce code duplication in ChangeColorspace
This commit decreases code duplication in the
parameter parsing of
augmenters.color.ChangeColorspace by using
the parameter handling functions in
parameters.py.

Additionally, the alpha parameter can now
be a list, which will be interpreted as a
Choice.

Additionally, the documentation of
Grayscale was updated in accordance to the
changes to ChangeColorspace.