Fix BG-augmenter returning original batches
The BackgroundAugmenter had a bug that would
lead the original (unaugmented) batches to be
returned instead of the augmented ones.
This commit fixes that bug.
The bug did not cause problems due to a
seperate bug that led to batches not being
copied during augmentation, but instead being
changed in-place.