Split test.py into multiple test_*.py files
test.py became too large, so it is now split
into multiple files, one for each file in imgaug/.
The old test.py file can still be executed in the
same way as before and will automatically call
the test_*.py files.
This change does not yet switch to unittest, which
will be done in another commit.