Make hooks default to None in augment_images()
hooks has previously defaulted to a new instance of
HooksImages in case it was set to None when calling
augment_images(). This is now removed and hooks
keeps the value None in these cases.
This should improve performance marginally by saving
a few hooks function calls and makes it clearer whether
the user has provided a hooks instance. In cases
where such an instance was not provided, steps related
to hooks may be skipped. This will enable to save
copies of images in child calls when no hooks were
specificied (i.e. switch to full in-place augmentation),
which should lead to a sizeable performance improvement.