Improve Uniform to return minimal float dtype
This patch changes parameters.Uniform(a, b)
so that it always returns the minimal float dtype
that still covers a and b.
This is assumed to improve performance for
following tasks that have to work with large
numbers of samples from Uniform, but e.g. don't
want to calculate min/max values.