Change docstrings from multiline datatypes to singleline datatypes
This patch changes how the datatypes of parameter are formatted
in docstrings.

For some datatype definitions of parameters in docstrings, the
different allowed datatypes were split over multiple lines to
make them easier to read and to avoid breaking the line length
limit. This is now changed so that these datatype definitions
always only consist of a single line, even if that severely
breaks the line length limit.

This is often horrendously ugly and very hard to read.
Unfortunately it seems to be the only format that is both
understood by Sphinx (when generating ReadTheDocs pages)
and PyCharm. Various ways seem to exist that make multiline
datatype definitions work in either of these cases, but
none that works in both cases. That is why it is now changed
to singleline definitions.