Remove 'dont import from this module' messages
The docstring of each module in imgaug.augmenters
previously included a suggestion to not directly
import from that module, but instead use
imgaug.augmenters.<AugmenterName>. That was due
to the categorization being still unstable.

As the categorization has now been fairly stable
for a long time, the suggestion is removed with
this patch.