Fix inaccuracy in PerspectiveTransformation
Fix an inaccuracy in `PerspectiveTransform` that can lead to
slightly misaligned transformations between images and
coordinate-based augmentables (e.g. bounding boxes). The problem
is more significant the smaller the images and larger the
`scale` values. It is also worsened by using `fit_output`.
The inaccuracy is caused by using H-1 and W-1 instead of
H and W.