Change **old_kwargs magic to explicit parameters (#595)
The parameter names of class `Augmenter` were recently changed
and the old names were caught by `**old_kwargs`. This was
convenient, but is changed in this patch to be explicitly
`random_state="deprecated", deterministic="deprecated"`.
That is because the `**old_kwargs` magic has the disadvantage
of not necessarily being fully understood by IDEs. These then
cannot detect (and warn) about invalid parameters anymore.
Without the magic parameter they can warn about these,
hence it is the preferred choice here.