Improve convolution filters
This patch reworks the backend of all convolutional
filters. It extracts the convolution logic out of
`Convolve` and moves it into the new function
`imgaug.augmenters.convolutional.convolve_()` (with
non-in-place version `convolve()`).

The logic is also reworked so that fewer convolution
function calls and more in-place modification is
used. This should lead to an improved performance.

These changes also affect `Sharpen`, `Emboss`,
`EdgeDetect`, `DirectedEdgeDetect` and `MotionBlur`.

Add functions:
* `imgaug.augmenters.convolutional.convolve_()`
* `imgaug.augmenters.convolutional.convolve()`