Refactor Affine
This patch refactors affine to make the code more
readable and change the matrix generation routine to
a numpy-based one. It also merges the matrix
generation of `Affine` and `pillike.Affine` and lays
the foundation for adding a `center` parameter to
`Affine`.

This patch also changes the shear mechanic in
`Affine`. When shearing on the x-axis, the points at
the top are now only moved to the left/right and no
longer up/down. Previously, they were also slightly
moved up/down. (Analogous for the y-axis.)